small fixed on ticket.java

+++ Phase4/Ticket.java
@@ -40,8 +40,7 @@ public class Ticket
 		return price;
 	}
 
-	public double getQuantity(){
+	public int getQuantity(){
 		return quantity;
 	}
-
 }
